backend/flask/requirements.txt
flask
flask-cors
flask-pymongo
requests
pymongo
gunicorn
flask-talisman
flask-limiter
pyjwt
